+++ Cmdlets/Get-ChildItem/Different-FileInfo-ToString/Test-1-different-ToString.ps1
@@ -6,6 +6,7 @@ $results = New-Object System.Collections.Specialized.OrderedDictionary
 $results.strings1 = [string[]](Get-ChildItem . -Include *)
 
 # this command gets just file names
+# (fixed in v6.1.0-preview.3)
 $results.strings2 = [string[]](Get-ChildItem .)
 
 # output